solved-python-easy-2 #1505

+++ tasks/python/easy/number_sum.py
@@ -3,4 +3,4 @@
 # TODO: Write a function that returns the sum of two numbers
 
 def twoSum(x,y):
-    return int(x)+int(y)
+    return float(x)+float(y)
